@@ -3,58 +3,73 @@
     "type": "project",
     "license": "MIT",
     "description": "A skeleton to start a new Symfony website",
-    "minimum-stability": "dev",
+    "minimum-stability": "stable",
     "require": {
-        "php": ">=7.1.3",
+        "php": ">=8.3",
         "ext-ctype": "*",
         "ext-iconv": "*",
-        "composer/package-versions-deprecated": "1.x-dev",
-        "doctrine/annotations": "^1.0",
-        "doctrine/doctrine-bundle": "^2",
-        "doctrine/doctrine-migrations-bundle": "^2",
-        "doctrine/orm": "^2",
-        "easycorp/easyadmin-bundle": "^2.0@dev",
-        "phpdocumentor/reflection-docblock": "5.x-dev",
-        "sensio/framework-extra-bundle": "^5.4@dev",
-        "symfony/apache-pack": "v1.0.1",
-        "symfony/asset": "4.4.*",
-        "symfony/console": "4.4.*",
-        "symfony/dotenv": "4.4.*",
-        "symfony/expression-language": "4.4.*",
-        "symfony/finder": "4.4.*",
-        "symfony/flex": "^1.1",
-        "symfony/form": "4.4.*",
-        "symfony/framework-bundle": "4.4.*",
-        "symfony/monolog-bundle": "^3.1",
-        "symfony/process": "4.4.*",
-        "symfony/property-access": "4.4.*",
-        "symfony/property-info": "4.4.*",
-        "symfony/security-bundle": "4.4.*",
-        "symfony/serializer": "4.4.*",
-        "symfony/swiftmailer-bundle": "^3.1",
-        "symfony/translation": "4.4.*",
-        "symfony/twig-bundle": "4.4.*",
-        "symfony/validator": "4.4.*",
-        "symfony/web-link": "4.4.*",
-        "symfony/yaml": "4.4.*"
+        "composer/package-versions-deprecated": "^1.11.99.5",
+        "doctrine/doctrine-bundle": "^2.11.2",
+        "doctrine/doctrine-migrations-bundle": "^3.3.0",
+        "doctrine/orm": "^2.18.0",
+        "phpdocumentor/reflection-docblock": "^5.3.0",
+        "symfony/asset": "7.0.*",
+        "symfony/console": "7.0.*",
+        "symfony/dotenv": "7.0.*",
+        "symfony/expression-language": "7.0.*",
+        "symfony/finder": "7.0.*",
+        "symfony/flex": "^2.4.4",
+        "symfony/form": "7.0.*",
+        "symfony/framework-bundle": "7.0.*",
+        "symfony/monolog-bundle": "^3.10",
+        "symfony/process": "7.0.*",
+        "symfony/property-access": "7.0.*",
+        "symfony/property-info": "7.0.*",
+        "symfony/runtime": "7.0.*",
+        "symfony/security-bundle": "7.0.*",
+        "symfony/serializer": "7.0.*",
+        "symfony/translation": "7.0.*",
+        "symfony/twig-bundle": "7.0.*",
+        "symfony/validator": "7.0.*",
+        "symfony/web-link": "7.0.*",
+        "symfony/yaml": "7.0.*"
     },
     "require-dev": {
-        "doctrine/doctrine-fixtures-bundle": "^3.3@dev",
-        "phpunit/phpunit": "10.0.x-dev",
-        "symfony/browser-kit": "4.4.*",
-        "symfony/css-selector": "4.4.*",
-        "symfony/debug-bundle": "4.4.*",
-        "symfony/maker-bundle": "^1.0",
-        "symfony/phpunit-bridge": "4.4.*",
-        "symfony/stopwatch": "4.4.*",
-        "symfony/web-profiler-bundle": "4.4.*",
-        "symfony/web-server-bundle": "4.4.*"
+        "dama/doctrine-test-bundle": "^8.0",
+        "dbrekelmans/bdi": "^1.1",
+        "doctrine/doctrine-fixtures-bundle": "^3.5.1",
+        "phpstan/extension-installer": "^1.3.1",
+        "phpstan/phpstan": "^1.10.57",
+        "phpstan/phpstan-deprecation-rules": "^1.1.4",
+        "phpstan/phpstan-doctrine": "^1.3.61",
+        "phpstan/phpstan-phpunit": "^1.3.15",
+        "phpstan/phpstan-symfony": "^1.3.7",
+        "phpunit/phpunit": "^9.6.16",
+        "qossmic/deptrac-shim": "^1.0.2",
+        "rector/rector": "^1.0.0",
+        "symfony/browser-kit": "7.0.*",
+        "symfony/css-selector": "7.0.*",
+        "symfony/debug-bundle": "7.0.*",
+        "symfony/maker-bundle": "^1.54.0",
+        "symfony/panther": "^2.1",
+        "symfony/phpunit-bridge": "7.0.*",
+        "symfony/stopwatch": "7.0.*",
+        "symfony/web-profiler-bundle": "7.0.*",
+        "symplify/easy-coding-standard": "^12.1.10"
     },
     "config": {
         "preferred-install": {
             "*": "dist"
         },
-        "sort-packages": true
+        "sort-packages": true,
+        "allow-plugins": {
+            "symfony/flex": true,
+            "phpstan/extension-installer": true,
+            "symfony/runtime": true
+        },
+        "platform": {
+            "php": "8.3"
+        }
     },
     "autoload": {
         "psr-4": {
