@@ -1,24 +1,25 @@
 <?php
 
-
 namespace App\Controller;
 
-use App\Entity\Career;
+use App\Repository\CareerRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Attribute\Route;
 
 class CareerController extends AbstractController
 {
-    /**
-     * @Route("/careers", name="careers")
-     */
-    public function index()
+    public function __construct(
+        private readonly CareerRepository $careerRepository,
+    ) {
+    }
+
+    #[Route(path: '/careers', name: 'careers')]
+    public function index(): Response
     {
-        $em = $this->getDoctrine()->getManager();
-        $careers = $em->getRepository(Career::class)->findAll();
         return $this->render('careers/index.html.twig', [
             'controller_name' => 'CareerController',
-            'careers' => $careers
+            'careers' => $this->careerRepository->findAll(),
         ]);
     }
 }
