@@ -2,22 +2,23 @@
 
 namespace App\Controller\Offer;
 
-use App\Entity\Destination;
+use App\Repository\DestinationRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Attribute\Route;
 
 class DestinationController extends AbstractController
 {
-    /**
-     * @Route("/destinations", name="destinations")
-     */
-    public function index()
+    public function __construct(
+        private readonly DestinationRepository $destinationRepository,
+    ) {
+    }
+
+    #[Route(path: '/destinations', name: 'destinations')]
+    public function index(): Response
     {
-        $em = $this->getDoctrine()->getManager();
-        $destinations = $em->getRepository(Destination::class)->findAll();
         return $this->render('destination/index.html.twig', [
-            'controller_name' => 'DestinationController',
-            'destinations' => $destinations
+            'destinations' => $this->destinationRepository->findAll(),
         ]);
     }
 }
