@@ -1,159 +1,179 @@
 <?php
 
-
 namespace App\Controller\Offer;
 
 use App\Entity\BookingOffer;
 use App\Entity\BookingOfferType;
 use App\Entity\Destination;
 use App\Entity\Reservation;
+use App\Entity\User;
 use App\Form\BookingOfferFiltersType;
+use App\Form\ConfirmReservationType;
 use App\Form\ReservationStartType;
+use App\Repository\BookingOfferRepository;
+use App\Repository\BookingOfferTypeRepository;
+use App\Repository\DestinationRepository;
+use App\Reservation\ReservationFactory;
 use App\Service\BookingOfferService;
-use DateTime;
-use Exception;
+use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Finder\Finder;
-use App\Form\ConfirmReservationType;
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Http\Attribute\CurrentUser;
 
-/**
- * @Route("/offer", name="offer_")
- */
+#[Route(path: '/offer', name: 'offer_')]
 class OfferController extends AbstractController
 {
+    public function __construct(
+        private readonly BookingOfferRepository $bookingOfferRepository,
+        private readonly BookingOfferTypeRepository $bookingOfferTypeRepository,
+        private readonly ReservationFactory $reservationFactory,
+        private readonly DestinationRepository $destinationRepository,
+        private readonly EntityManagerInterface $entityManager,
+    ) {
+    }
+
     /**
-     * @Route("/browse", name="browse")
-     * @param Request $request
-     * @param BookingOfferService $offerService
-     * @return Response
-     * @throws Exception
+     * @throws \Exception
      */
-    public function displayOfferList(Request $request, BookingOfferService $offerService)
+    #[Route(path: '/browse', name: 'browse')]
+    public function displayOfferList(Request $request, BookingOfferService $offerService): Response
     {
         $bookingOffer = new BookingOffer();
-        if ($request->query->get('booking_offer_search')) {
+        if ($request->query->has('booking_offer_search')) {
             $offers = $this->getOffersBasedOnRequestQuery($offerService, $request, $bookingOffer);
         } elseif ($request->query->get('offerType')) {
             $offers = $this->getOffersBasedOnOfferType($offerService, $request->query->get('offerType'));
         } elseif ($request->query->get('destinationName')) {
             $offers = $this->getOffersBasedOnDestination($offerService, $request->query->get('destinationName'));
-            $queryDestination = $this->getDoctrine()->getRepository(Destination::class)
-                ->findOneBy(['destinationName' => $request->query->get('destinationName')]);
+            $queryDestination = $this->destinationRepository->findOneBy(['destinationName' => $request->query->get('destinationName')]);
             $bookingOffer->setDestination($queryDestination);
         } else {
             $offers = $offerService->findOffers();
         }
-        $allOffers = $this->getDoctrine()->getRepository(BookingOffer::class)->findAll();
-        foreach ($allOffers as $offer) {
-            $departureSpots[$offer->getDepartureSpot()] = $offer->getDepartureSpot();
-        }
+
         $filtersForm = $this->createForm(BookingOfferFiltersType::class, $bookingOffer, [
             'method' => 'GET',
-            'departureSpots' => $departureSpots
         ]);
+
         if ($request->query->get('offerType')) {
-            $fetchedType = $this->getDoctrine()->getRepository(BookingOfferType::class)
-                ->findOneBy(['typeName' => $request->query->get('offerType')]);
-            if ($fetchedType) {
-                $typeId = $fetchedType->getId();
-                $filtersForm->get('offerTypes')->get("$typeId")->setData(true);
+            $fetchedType = $this->bookingOfferTypeRepository->findOneBy(['typeName' => $request->query->get('offerType')]);
+            if (null !== $fetchedType) {
+                $filtersForm->get('offerTypes')->get($fetchedType->getId())->setData(true);
             }
         }
+
         $filtersForm->handleRequest($request);
         if ($filtersForm->isSubmitted() && $filtersForm->isValid()) {
             $offers = $this->getOffersBasedOnFormSubmission($offerService, $filtersForm, $bookingOffer);
         }
+
         return $this->render('offer/browser.html.twig', [
             'offers' => $offers,
             'parameters' => $request->attributes->all(),
-            'filtersForm' => $filtersForm
+            'offer_filters_form' => $filtersForm->createView(),
         ]);
     }
 
-    /**
-     * @Route ("/reservationSummary/{offerId}/adults/{adultNumber}/children/{childNumber}", name="reservationSummary")
-     * @param Request $request
-     * @param int $offerId
-     * @param int $adultNumber
-     * @param int $childNumber
-     * @return Response
-     */
-    public function displayReservationSummary(Request $request, int $offerId, int $adultNumber, int $childNumber)
-    {
-        $reservation = new Reservation();
-        $offer = $this->getDoctrine()->getRepository(BookingOffer::class)->find($offerId);
-        $reservation->setBookingOffer($offer);
-        $reservation->setAdultNumber($adultNumber);
-        $reservation->setChildNumber($childNumber);
-        $reservation->setBankTransferTitle();
-        $reservation->setTotalCost($this->getReservationTotalCost($reservation));
-        $reservation->setUser($this->getUser());
+    #[Route(path: '/reservationSummary/{offerId}/adults/{adultNumber}/children/{childNumber}', name: 'reservationSummary')]
+    public function displayReservationSummary(
+        Request $request,
+        #[CurrentUser] User $user,
+        int $offerId,
+        int $adultNumber,
+        int $childNumber,
+    ): Response {
+        // TODO: handle expired offers
+        $offer = $this->bookingOfferRepository->find($offerId);
+
+        if (!$offer instanceof BookingOffer) {
+            throw $this->createNotFoundException(\sprintf('Unable to find booking offer identified by ID `%s`.', $offerId));
+        }
+
+        $reservation = $this->reservationFactory->create($user, $offer, $adultNumber, $childNumber);
+
         $form = $this->createForm(ConfirmReservationType::class, $reservation);
         $form->handleRequest($request);
+
         if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $reservation->setDateOfBooking(new DateTime('NOW'));
-            $reservation->setIsPaidFor(false);
-            $em->persist($reservation);
-            $em->flush();
-            return $this->redirectToRoute("reservations");
+            $reservation->confirm();
+            $this->entityManager->persist($reservation);
+            $this->entityManager->flush();
+
+            return $this->redirectToRoute('reservations');
         }
+
         return $this->render('offer/reservation_summary.html.twig', [
             'reservation' => $reservation,
-            'form' => $form
+            'form' => $form->createView(),
         ]);
     }
 
-    /**
-     * @Route ("/{id}", name="single")
-     * @param Request $request
-     * @param int $id
-     * @return Response
-     */
-    public function displayOffer(Request $request, $id)
+    #[Route(path: '/{id}', name: 'single')]
+    public function displayOffer(Request $request, int $id): Response
     {
-        $offer = $this->getDoctrine()->getRepository(BookingOffer::class)->findOffer($id);
+        // TODO: handle not found or expired offer
+        $offer = $this->bookingOfferRepository->findOffer($id);
+
         $finder = new Finder();
         $finder->files()->in($offer->getPhotosDirectory());
+
         $photosCount = 0;
         if ($finder->hasResults()) {
             $photosCount = $finder->count();
         }
+
         $reservation = new Reservation();
         $reservation->setBookingOffer($offer);
+
         $reservationForm = $this->createForm(ReservationStartType::class, $reservation);
         $reservationForm->handleRequest($request);
         if ($reservationForm->isSubmitted() && $reservationForm->isValid()) {
             return $this->redirectToRoute('offer_reservationSummary', [
                 'offerId' => $id,
                 'adultNumber' => $reservation->getAdultNumber(),
-                'childNumber' => $reservation->getChildNumber()
+                'childNumber' => $reservation->getChildNumber(),
             ]);
         }
+
         return $this->render('offer/single_offer.html.twig', [
             'offer' => $offer,
-            'reservationForm' => $reservationForm,
-            'photosCount' => $photosCount
+            'init_reservation_form' => $reservationForm->createView(),
+            'photosCount' => $photosCount,
         ]);
     }
 
-
-    private function getWellFormattedDate(string $date)
+    /**
+     * TODO: extract from controller code.
+     */
+    private function getWellFormattedDate(string $date): \DateTimeInterface
     {
-        $reformatted = explode('/', $date);
-        $reformatted = $reformatted[2] . '/' . $reformatted[1] . '/' . $reformatted[0];
-        return new DateTime($reformatted);
+        $reformatted = \explode('/', $date);
+        $reformatted = $reformatted[2].'/'.$reformatted[1].'/'.$reformatted[0];
+
+        return new \DateTime($reformatted);
     }
 
-    private function getOffersBasedOnFormSubmission(BookingOfferService $offerService, FormInterface $filtersForm, BookingOffer $bookingOffer)
-    {
+    /**
+     * @return iterable<int, BookingOffer>
+     */
+    private function getOffersBasedOnFormSubmission(
+        BookingOfferService $offerService,
+        FormInterface $filtersForm,
+        BookingOffer $bookingOffer,
+    ): iterable {
         $priceMin = $filtersForm->get('priceMin')->getData();
+        \assert(\is_numeric($priceMin) || null === $priceMin);
+
         $priceMax = $filtersForm->get('priceMax')->getData();
+        \assert(\is_numeric($priceMax) || null === $priceMax);
+
+        /** @var BookingOfferType[]|null $offerTypes */
         $offerTypes = $filtersForm->get('offerTypes')->getData();
+
         return $offerService->findOffers(
             $bookingOffer->getDepartureSpot(),
             $bookingOffer->getDestination(),
@@ -161,57 +181,61 @@ class OfferController extends AbstractController
             $bookingOffer->getComebackDate(),
             $priceMin,
             $priceMax,
-            $offerTypes
+            $offerTypes,
         );
     }
 
-    private function getOffersBasedOnRequestQuery(BookingOfferService $offerService, Request $request, BookingOffer $bookingOffer)
-    {
-        $requestParams = $request->query->get('booking_offer_search');
+    /**
+     * @return iterable<int, BookingOffer>
+     */
+    private function getOffersBasedOnRequestQuery(
+        BookingOfferService $offerService,
+        Request $request,
+        BookingOffer $bookingOffer,
+    ): iterable {
+        $requestParams = $request->query->all('booking_offer_search');
         $departureSpot = $requestParams['departureSpot'] ?? null;
         $destination = $requestParams['destination'] ?? null;
         $departureDate = $requestParams['departureDate'] ?? null;
         $comebackDate = $requestParams['comebackDate'] ?? null;
-        if ($departureDate != null) {
+        if (null != $departureDate) {
             $bookingOffer->setDepartureDate($this->getWellFormattedDate($departureDate));
         }
-        if ($comebackDate != null) {
+
+        if (null != $comebackDate) {
             $bookingOffer->setComebackDate($this->getWellFormattedDate($comebackDate));
         }
+
         $bookingOffer->setDepartureSpot($departureSpot);
-        $bookingOffer->setDestination($this->getDoctrine()->getRepository(Destination::class)->find($destination));
+        $bookingOffer->setDestination($this->destinationRepository->find($destination));
+
         return $offerService->findOffers($departureSpot,
             $destination,
             $bookingOffer->getDepartureDate(),
             $bookingOffer->getComebackDate());
     }
 
-    private function getOffersBasedOnOfferType(BookingOfferService $offerService, string $offerTypeName)
+    /**
+     * @return iterable<int, BookingOffer>
+     */
+    private function getOffersBasedOnOfferType(BookingOfferService $offerService, string $offerTypeName): iterable
     {
-        $offerType = $this->getDoctrine()->getRepository(BookingOfferType::class)->findOneBy(['typeName' => $offerTypeName]);
-        if ($offerType != null) {
-            $offers = $offerService->findOffers(null, null, null, null, null, null, [$offerType]);
-        } else {
-            $offers = null;
-        }
-        return $offers;
+        $offerType = $this->bookingOfferTypeRepository->findOneBy(['typeName' => $offerTypeName]);
+
+        return $offerType instanceof BookingOfferType
+            ? $offerService->findOffers(bookingOfferTypes: [$offerType])
+            : [];
     }
 
-    private function getOffersBasedOnDestination(BookingOfferService $offerService, string $destinationName)
+    /**
+     * @return iterable<int, BookingOffer>
+     */
+    private function getOffersBasedOnDestination(BookingOfferService $offerService, string $destinationName): iterable
     {
-        $destination = $this->getDoctrine()->getRepository(Destination::class)->findOneBy(['destinationName' => $destinationName]);
-        if ($destination != null) {
-            $offers = $offerService->findOffers(null, $destination, null, null, null, null, null);
-        } else {
-            $offers = null;
-        }
-        return $offers;
-    }
+        $destination = $this->destinationRepository->findOneBy(['destinationName' => $destinationName]);
 
-    private function getReservationTotalCost($reservation)
-    {
-        $adultPrice = $reservation->getBookingOffer()->getOfferPrice();
-        $childPrice = $reservation->getBookingOffer()->getChildPrice();
-        return $reservation->getAdultNumber() * $adultPrice + $reservation->getChildNumber() * $childPrice;
+        return $destination instanceof Destination
+            ? $offerService->findOffers(destination: $destination)
+            : [];
     }
 }
