@@ -3,24 +3,19 @@
 namespace App\Entity;
 
 use App\Repository\NewsletterRepository;
+use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
 
-/**
- * @ORM\Entity(repositoryClass=NewsletterRepository::class)
- */
+#[ORM\Entity(repositoryClass: NewsletterRepository::class)]
 class Newsletter
 {
-    /**
-     * @ORM\Id
-     * @ORM\GeneratedValue
-     * @ORM\Column(type="integer")
-     */
-    private $id;
+    #[ORM\Id]
+    #[ORM\GeneratedValue]
+    #[ORM\Column(type: Types::INTEGER)]
+    private ?int $id = null;
 
-    /**
-     * @ORM\Column(type="string", length=255)
-     */
-    private $email;
+    #[ORM\Column(type: Types::STRING, length: 255)]
+    private ?string $email = null;
 
     public function getId(): ?int
     {
