@@ -1,50 +1,62 @@
 <?php
 
-
 namespace App\Controller\UsersData;
 
-use App\Entity\Reservation;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use App\Entity\CustomersRating;
-use App\Entity\BookingOffer;
+use App\Entity\User;
 use App\Form\RateOfferType;
+use App\Repository\ReservationRepository;
+use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Http\Attribute\CurrentUser;
 
 class RateOfferController extends AbstractController
 {
-    /**
-     * @Route ("rateOffer/{reservationId}", name="rateOffer")
-     * @param Request $request
-     * @param int $reservationId
-     * @return Response
-     */
-    public function displayRateOfferForm(Request $request, int $reservationId)
-    {
-        if ((isset($_SESSION['display_rate_offer'][$reservationId]) && $_SESSION['display_rate_offer'][$reservationId] === TRUE) || !empty($request->request->all())){
-            $_SESSION['display_rate_offer'][$reservationId] = FALSE;
-            $reservation = $this->getDoctrine()->getRepository(Reservation::class)->find($reservationId);
+    public function __construct(
+        private readonly ReservationRepository $reservationRepository,
+        private readonly EntityManagerInterface $entityManager,
+    ) {
+    }
+
+    #[Route(path: 'rateOffer/{reservationId}', name: 'rateOffer')]
+    public function displayRateOfferForm(
+        Request $request,
+        int $reservationId,
+        #[CurrentUser] User $user,
+    ): Response {
+        $session = $request->getSession();
+
+        /** @var array<int, bool> $displayRateOffers */
+        $displayRateOffers = $session->get('display_rate_offer', []);
+        $displayRateOffer = $displayRateOffers[$reservationId] ?? false;
+
+        if ($displayRateOffer || [] !== $request->request->all()) {
+            $session->set('display_rate_offer', \array_merge($displayRateOffers, [$reservationId => false]));
+
+            $reservation = $this->reservationRepository->find($reservationId);
             $offer = $reservation->getBookingOffer();
             $packageId = $offer->getPackageId();
             $customersRating = new CustomersRating();
             $customersRating->setPackage($packageId);
-            $user = $this->getUser();
             $customersRating->setUser($user);
             $rateOfferForm = $this->createForm(RateOfferType::class, $customersRating);
             $rateOfferForm->handleRequest($request);
             if ($rateOfferForm->isSubmitted() && $rateOfferForm->isValid()) {
-                $em = $this->getDoctrine()->getManager();
-                $em->persist($customersRating);
-                $em->flush();
-                return $this->redirectToRoute("reservations");
+                $this->entityManager->persist($customersRating);
+                $this->entityManager->flush();
+
+                return $this->redirectToRoute('reservations');
             }
+
             return $this->render('reservations/rateOffer.html.twig', [
                 'offer' => $offer,
-                'rateOfferForm' => $rateOfferForm
+                'rate_offer_form' => $rateOfferForm->createView(),
             ]);
-        } else {
-            return $this->redirectToRoute("reservations");
         }
+
+        return $this->redirectToRoute('reservations');
     }
 }
