@@ -1,51 +1,48 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\DataFixtures;
 
-use Doctrine\Bundle\FixturesBundle\Fixture;
 use App\Entity\BookingOfferType;
+use Doctrine\Bundle\FixturesBundle\Fixture;
 use Doctrine\Persistence\ObjectManager;
 
-class BookingOfferTypeFixtures extends Fixture
+final class BookingOfferTypeFixtures extends Fixture
 {
-    public const FIRST_MINUTE_REFERENCE = 'first_minute';
-    public const LAST_MINUTE_REFERENCE = 'last_minute';
-    public const ALL_INCLUSIVE_REFERENCE = 'all_inclusive';
-    public const FOR_CHILDREN_REFERENCE = 'for_children';
-    public const GROUP_TOURS_REFERENCE = 'group_tours';
-    public const CRUISES_REFERENCE = 'cruises';
+    public const string FIRST_MINUTE_REFERENCE = 'first_minute';
 
-    public function load(ObjectManager $manager)
+    public const string LAST_MINUTE_REFERENCE = 'last_minute';
+
+    public const string ALL_INCLUSIVE_REFERENCE = 'all_inclusive';
+
+    public const string FOR_CHILDREN_REFERENCE = 'for_children';
+
+    public const string GROUP_TOURS_REFERENCE = 'group_tours';
+
+    public const string CRUISES_REFERENCE = 'cruises';
+
+    /**
+     * @var array<string, string>
+     */
+    private const array NAME_MAP = [
+        self::FIRST_MINUTE_REFERENCE => 'First Minute',
+        self::LAST_MINUTE_REFERENCE => 'Last Minute',
+        self::ALL_INCLUSIVE_REFERENCE => 'All Inclusive',
+        self::FOR_CHILDREN_REFERENCE => 'For Children',
+        self::GROUP_TOURS_REFERENCE => 'Group Tours',
+        self::CRUISES_REFERENCE => 'Cruises',
+    ];
+
+    #[\Override]
+    public function load(ObjectManager $manager): void
     {
-        $offerType = new BookingOfferType();
-        $offerType->setTypeName('First Minute');
-        $this->addReference(self::FIRST_MINUTE_REFERENCE,$offerType);
-        $manager->persist($offerType);
+        foreach (self::NAME_MAP as $referenceKey => $name) {
+            $offerType = new BookingOfferType($name);
+            $manager->persist($offerType);
 
-        $offerType = new BookingOfferType();
-        $offerType->setTypeName('Last Minute');
-        $this->addReference(self::LAST_MINUTE_REFERENCE,$offerType);
-        $manager->persist($offerType);
-
-        $offerType = new BookingOfferType();
-        $offerType->setTypeName('All Inclusive');
-        $this->addReference(self::ALL_INCLUSIVE_REFERENCE,$offerType);
-        $manager->persist($offerType);
-
-        $offerType = new BookingOfferType();
-        $offerType->setTypeName('For Children');
-        $this->addReference(self::FOR_CHILDREN_REFERENCE,$offerType);
-        $manager->persist($offerType);
-
-        $offerType = new BookingOfferType();
-        $offerType->setTypeName('Group Tours');
-        $this->addReference(self::GROUP_TOURS_REFERENCE,$offerType);
-        $manager->persist($offerType);
-
-        $offerType = new BookingOfferType();
-        $offerType->setTypeName('Cruises');
-        $this->addReference(self::CRUISES_REFERENCE,$offerType);
-        $manager->persist($offerType);
+            $this->addReference($referenceKey, $offerType);
+        }
 
         $manager->flush();
     }
