@@ -1,40 +1,50 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\DataFixtures;
 
 use App\Entity\Career;
-use DateTime;
 use Doctrine\Bundle\FixturesBundle\Fixture;
 use Doctrine\Persistence\ObjectManager;
 
-class CareerFixture extends Fixture
+final class CareerFixture extends Fixture
 {
-
-    public function load(ObjectManager $manager)
+    #[\Override]
+    public function load(ObjectManager $manager): void
     {
         $career = $this->createCareer(
             'manager',
             'Team management',
             '2 years experience',
-            4500.00,
-            new DateTime('2020-11-22'),
-            new DateTime('2020-12-30')
+            '4500.00',
+            new \DateTime('-5 days'),
+            new \DateTime('+12 months')
         );
+
         $manager->persist($career);
+
         $career = $this->createCareer(
             'tour guide',
             'Tour guidance',
             'Student status',
-            2000.00,
-            new DateTime('2020-11-22'),
-            new DateTime('2020-12-30')
+            '2000.00',
+            new \DateTime('-12 days'),
+            new \DateTime('+8 months')
         );
+
         $manager->persist($career);
         $manager->flush();
     }
 
-    private function createCareer($jobTitle, $description, $requirements, $salary, $startDate, $endDate)
-    {
+    private function createCareer(
+        string $jobTitle,
+        string $description,
+        string $requirements,
+        string $salary,
+        \DateTime $startDate,
+        \DateTime $endDate,
+    ): Career {
         $career = new Career();
         $career->setJobTitle($jobTitle);
         $career->setDescription($description);
@@ -42,6 +52,7 @@ class CareerFixture extends Fixture
         $career->setSalary($salary);
         $career->setRecruitmentStartDate($startDate);
         $career->setRecruitmentEndDate($endDate);
+
         return $career;
     }
 }
