@@ -1,5 +1,6 @@
 <?php
 
+declare(strict_types=1);
 
 namespace App\DataFixtures;
 
@@ -7,19 +8,22 @@ use App\Entity\Newsletter;
 use Doctrine\Bundle\FixturesBundle\Fixture;
 use Doctrine\Persistence\ObjectManager;
 
-class NewsletterFixture extends Fixture
+final class NewsletterFixture extends Fixture
 {
-    public function load(ObjectManager $manager)
+    #[\Override]
+    public function load(ObjectManager $manager): void
     {
         $manager->persist($this->createNewsletter('john_cena@holidaydream.com'));
         $manager->persist($this->createNewsletter('marta@mazurkiewi.cz'));
 
         $manager->flush();
     }
-    private function createNewsletter($email)
+
+    private function createNewsletter(string $email): Newsletter
     {
         $newsletter = new Newsletter();
         $newsletter->setEmail($email);
+
         return $newsletter;
     }
 }
