@@ -1,101 +1,138 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\DataFixtures;
 
+use App\Entity\BookingOffer;
 use App\Entity\Reservation;
-use DateTime;
+use App\Entity\User;
+use App\Reservation\ReservationFactory;
 use Doctrine\Bundle\FixturesBundle\Fixture;
 use Doctrine\Common\DataFixtures\DependentFixtureInterface;
 use Doctrine\Persistence\ObjectManager;
 
-class ReservationFixture extends Fixture implements DependentFixtureInterface
+final class ReservationFixture extends Fixture implements DependentFixtureInterface
 {
+    public function __construct(
+        private readonly ReservationFactory $reservationFactory,
+    ) {
+    }
 
-    public function load(ObjectManager $manager)
+    #[\Override]
+    public function load(ObjectManager $manager): void
     {
-        $reservation = $this->createReservation(
-            $this->getReference(UserFixture::USER1_REFERENCE),
-            $this->getReference(BookingOfferFixture::BEIJING_REFERENCE),
+        $manager->persist($this->createReservation(
+            UserFixture::USER1_REFERENCE,
+            BookingOfferFixture::BEIJING_REFERENCE,
             1,
             0,
-            new DateTime('2020-11-22')
-        );
-        $manager->persist($reservation);
+            new \DateTime('2020-11-22'),
+        ));
 
-        $reservation = $this->createPaidReservation(
-            $this->getReference(UserFixture::USER1_REFERENCE),
-            $this->getReference(BookingOfferFixture::PATAGONIA_REFERENCE),
+        $manager->persist($this->createPaidReservation(
+            UserFixture::USER1_REFERENCE,
+            BookingOfferFixture::PATAGONIA_REFERENCE,
             2,
             2,
-            new DateTime('2020-11-22'),
-            new DateTime('2020-11-24')
-        );
-        $manager->persist($reservation);
+            new \DateTime('2020-11-22'),
+            new \DateTime('2020-11-24'),
+        ));
 
-        $reservation = $this->createPaidReservation(
-            $this->getReference(UserFixture::USER2_REFERENCE),
-            $this->getReference(BookingOfferFixture::MAFIOSO_REFERENCE1),
+        $manager->persist($this->createPaidReservation(
+            UserFixture::USER2_REFERENCE,
+            BookingOfferFixture::MAFIOSO_REFERENCE1,
             1,
             2,
-            new DateTime('2019-06-22'),
-            new DateTime('2019-06-24')
-        );
-        $manager->persist($reservation);
+            new \DateTime('2019-06-22'),
+            new \DateTime('2019-06-24'),
+        ));
 
-        $reservation = $this->createReservation(
-            $this->getReference(UserFixture::USER2_REFERENCE),
-            $this->getReference(BookingOfferFixture::PATAGONIA_REFERENCE),
+        $manager->persist($this->createReservation(
+            UserFixture::USER2_REFERENCE,
+            BookingOfferFixture::PATAGONIA_REFERENCE,
             5,
             0,
-            new DateTime('2020-11-23')
-        );
-        $manager->persist($reservation);
+            new \DateTime('2020-11-23'),
+        ));
 
-        $reservation = $this->createPaidReservation(
-            $this->getReference(UserFixture::USER2_REFERENCE),
-            $this->getReference(BookingOfferFixture::BEIJING_REFERENCE),
+        $manager->persist($this->createPaidReservation(
+            UserFixture::USER2_REFERENCE,
+            BookingOfferFixture::BEIJING_REFERENCE,
             3,
             2,
-            new DateTime('2020-11-22'),
-            new DateTime('2020-11-22')
-        );
-        $manager->persist($reservation);
+            new \DateTime('2020-11-22'),
+            new \DateTime('2020-11-22'),
+        ));
 
-        $reservation = $this->createPaidReservation(
-            $this->getReference(UserFixture::USER2_REFERENCE),
-            $this->getReference(BookingOfferFixture::MAHARAJA_REFERENCE1),
+        $manager->persist($this->createPaidReservation(
+            UserFixture::USER2_REFERENCE,
+            BookingOfferFixture::MAHARAJA_REFERENCE1,
             2,
             0,
-            new DateTime('2020-04-22'),
-            new DateTime('2020-04-24')
-        );
-        $manager->persist($reservation);
+            new \DateTime('2020-04-22'),
+            new \DateTime('2020-04-24'),
+        ));
 
         $manager->flush();
     }
 
-    private function createReservation($user, $bookingOffer, $adultNumber, $childNumber, $dateOfBooking, $paid = false): Reservation
-    {
-        $reservation = new Reservation();
-        $reservation->setUser($user);
-        $reservation->setBookingOffer($bookingOffer);
-        $reservation->setAdultNumber($adultNumber);
-        $reservation->setChildNumber($childNumber);
-        $reservation->setDateOfBooking($dateOfBooking);
-        $reservation->setIsPaidFor($paid);
-        $reservation->setBankTransferTitle();
+    private function createReservation(
+        string $user,
+        string $bookingOffer,
+        int $adultNumber,
+        int $childNumber,
+        \DateTime $dateOfBooking,
+        bool $paid = false,
+    ): Reservation {
+        $reservation = $this->reservationFactory->create(
+            customer: $this->getUser($user),
+            bookingOffer: $this->getBookingOffer($bookingOffer),
+            numberOfAdults: $adultNumber,
+            numberOfChildren: $childNumber,
+        );
+
+        $reservation->confirm($dateOfBooking, $paid ? $reservation->getTotalCost() : null);
+
         return $reservation;
     }
 
-    private function createPaidReservation($user, $bookingOffer, $adultNumber, $childNumber, $dateOfBooking, $bankTransferDate): Reservation
-    {
-        $reservation = $this->createReservation($user, $bookingOffer, $adultNumber, $childNumber, $dateOfBooking, true);
+    private function createPaidReservation(
+        string $user,
+        string $bookingOffer,
+        int $adultNumber,
+        int $childNumber,
+        \DateTime $dateOfBooking,
+        \DateTime $bankTransferDate,
+    ): Reservation {
+        $reservation = $this->createReservation($user, $bookingOffer, $adultNumber, $childNumber, $dateOfBooking, paid: true);
         $reservation->setBankTransferDate($bankTransferDate);
+
         return $reservation;
     }
 
-    public function getDependencies()
+    /**
+     * @return class-string[]
+     */
+    #[\Override]
+    public function getDependencies(): array
     {
         return [UserFixture::class, BookingOfferFixture::class];
+    }
+
+    private function getUser(string $referenceKey): User
+    {
+        $user = $this->getReference($referenceKey);
+        \assert($user instanceof User);
+
+        return $user;
+    }
+
+    private function getBookingOffer(string $referenceKey): BookingOffer
+    {
+        $bookingOffer = $this->getReference($referenceKey);
+        \assert($bookingOffer instanceof BookingOffer);
+
+        return $bookingOffer;
     }
 }
