@@ -1,41 +1,43 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\DataFixtures;
 
+use App\Entity\User;
 use Doctrine\Bundle\FixturesBundle\Fixture;
 use Doctrine\Persistence\ObjectManager;
-use App\Entity\User;
-use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
+use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
 
-class UserFixture extends Fixture
+final class UserFixture extends Fixture
 {
-    public const USER1_REFERENCE = 'user1';
-    public const USER2_REFERENCE = 'user2';
-    public const ADMIN_REFERENCE = 'admin';
+    public const string USER1_REFERENCE = 'user1';
 
-    private $passwordEncoder;
+    public const string USER2_REFERENCE = 'user2';
 
-    public function __construct(UserPasswordEncoderInterface $passwordEncoder)
-    {
-        $this->passwordEncoder = $passwordEncoder;
+    public const string ADMIN_REFERENCE = 'admin';
+
+    public function __construct(
+        private readonly UserPasswordHasherInterface $passwordHasher,
+    ) {
     }
 
-    public function load(ObjectManager $manager)
+    #[\Override]
+    public function load(ObjectManager $manager): void
     {
-        $user1 = $this->createUser('Jan', 'Kowalski', 'jan_kowalski@dreamholiday.com');
-        $this->addReference(self::USER1_REFERENCE,$user1);
+        $user1 = $this->createUser('John', 'Doe', 'john.doe@dreamholiday.com');
+        $this->addReference(self::USER1_REFERENCE, $user1);
         $manager->persist($user1);
 
-        $user2 = $this->createUser('John', 'Cena', 'john_cena@holidaydream.com');
-        $this->addReference(self::USER2_REFERENCE,$user2);
+        $user2 = $this->createUser('Alice', 'Smith', 'alice.smith@holidaydream.com');
+        $this->addReference(self::USER2_REFERENCE, $user2);
         $manager->persist($user2);
 
-        $admin = $this->createAdmin('Jacob', 'Ćwikowski', 'kcwikowski007@gmail.com');
-        $this->addReference(self::ADMIN_REFERENCE,$admin);
+        $admin = $this->createAdmin('Super', 'Admin', 'super.admin@holidaydream.com');
+        $this->addReference(self::ADMIN_REFERENCE, $admin);
         $manager->persist($admin);
 
         $manager->flush();
-
     }
 
     private function createUser(string $firstName, string $lastName, string $email): User
@@ -44,8 +46,9 @@ class UserFixture extends Fixture
         $user->setFirstName($firstName);
         $user->setLastName($lastName);
         $user->setEmail($email);
-        $user->setPassword($this->passwordEncoder->encodePassword($user, 'password'));
+        $user->setPassword($this->passwordHasher->hashPassword($user, 'password'));
         $user->setRegistrationDate(new \DateTime('now'));
+
         return $user;
     }
 
@@ -53,6 +56,7 @@ class UserFixture extends Fixture
     {
         $admin = $this->createUser($firstName, $lastName, $email);
         $admin->addRole('ROLE_ADMIN');
+
         return $admin;
     }
 }
