@@ -1,8 +1,6 @@
 <?php
 
-
 namespace App\Form;
-
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\PasswordType;
@@ -12,6 +10,7 @@ use Symfony\Component\OptionsResolver\OptionsResolver;
 
 class AuthType extends AbstractType
 {
+    #[\Override]
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
@@ -19,17 +18,19 @@ class AuthType extends AbstractType
                 'required' => true,
                 'label' => 'Current Password',
                 'attr' => [
-                    'class' => 'form-control'
-                ]
+                    'class' => 'form-control',
+                ],
             ])
             ->add('submit', SubmitType::class, [
                 'label' => 'Sign in',
                 'attr' => [
-                    'class' => 'btn btn-login'
-                ]
-            ]);
+                    'class' => 'btn btn-login',
+                ],
+            ])
+        ;
     }
 
+    #[\Override]
     public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver->setDefaults([
