@@ -15,28 +15,29 @@ use Symfony\Component\OptionsResolver\OptionsResolver;
 
 class BookingOfferSearchType extends AbstractType
 {
-    public function buildForm(FormBuilderInterface $builder, array $options)
+    #[\Override]
+    public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
             ->add('departureDate', TextType::class, [
                 'attr' => [
                     'class' => 'form-control datepicker',
-                    'placeholder' => "Depart Date"
+                    'placeholder' => 'Depart Date',
                 ],
                 'label_attr' => [
-                    'class' => 'sr-only'
+                    'class' => 'sr-only',
                 ],
-                'required' => false
+                'required' => false,
             ])
             ->add('comebackDate', TextType::class, [
                 'attr' => [
                     'class' => 'form-control datepicker',
-                    'placeholder' => 'Return Date'
+                    'placeholder' => 'Return Date',
                 ],
                 'label_attr' => [
-                    'class' => 'sr-only'
+                    'class' => 'sr-only',
                 ],
-                'required' => false
+                'required' => false,
             ])
             ->add('departureSpot', ChoiceType::class, [
                 'choices' => $options['departureSpots'],
@@ -45,65 +46,75 @@ class BookingOfferSearchType extends AbstractType
                     'class' => 'form-control',
                 ],
                 'label_attr' => [
-                    'class' => 'sr-only'
+                    'class' => 'sr-only',
                 ],
-                'required' => false
+                'required' => false,
             ])
             ->add('destination', EntityType::class, [
                 'choices' => $options['destinations'],
                 'class' => Destination::class,
                 'attr' => [
-                    'class' => 'form-control'
+                    'class' => 'form-control',
                 ],
                 'placeholder' => 'To',
                 'label' => 'To',
                 'label_attr' => [
-                    'class' => 'sr-only'
+                    'class' => 'sr-only',
                 ],
-                'required' => false
+                'required' => false,
             ])
             ->add('submit', SubmitType::class, [
-                'label' => false
-            ]);
+                'label' => false,
+            ])
+        ;
 
-            $builder->get('departureDate')->addModelTransformer( new CallbackTransformer(
-                function ($date) {
-                    if($date!=null)
-                        return $date->format('d/m/Y');
-                    return null;
-                },
-                function ($date) {
-                    if($date!=null) {
-                        $date = explode('/',$date);
-                        $date = $date[2] . '/' . $date[1] . '/' . $date[0];
-                        return new \DateTime($date);
-                    }
-                    return null;
+        $builder->get('departureDate')->addModelTransformer(new CallbackTransformer(
+            static function ($date) {
+                if (null != $date) {
+                    return $date->format('d/m/Y');
                 }
-            ));
-            $builder->get('comebackDate')->addModelTransformer( new CallbackTransformer(
-                function ($date) {
-                    if($date!=null)
-                        return $date->format('d/m/Y');
-                    return null;
-                },
-                function ($date) {
-                    if($date!=null) {
-                        $date = explode('/',$date);
-                        $date = $date[2] . '/' . $date[1] . '/' . $date[0];
-                        return new \DateTime($date);
-                    }
-                    return null;
+
+                return null;
+            },
+            static function ($date): ?\DateTime {
+                if (null != $date) {
+                    $date = explode('/', $date);
+                    $date = $date[2].'/'.$date[1].'/'.$date[0];
+
+                    return new \DateTime($date);
                 }
-            ));
+
+                return null;
+            }
+        ));
+        $builder->get('comebackDate')->addModelTransformer(new CallbackTransformer(
+            static function ($date) {
+                if (null != $date) {
+                    return $date->format('d/m/Y');
+                }
+
+                return null;
+            },
+            static function ($date): ?\DateTime {
+                if (null != $date) {
+                    $date = explode('/', $date);
+                    $date = $date[2].'/'.$date[1].'/'.$date[0];
+
+                    return new \DateTime($date);
+                }
+
+                return null;
+            }
+        ));
     }
 
-    public function configureOptions(OptionsResolver $resolver)
+    #[\Override]
+    public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver->setDefaults([
             'data_class' => BookingOffer::class,
             'destinations' => null,
-            'departureSpots' => null
+            'departureSpots' => null,
         ]);
     }
 }
