@@ -1,49 +1,51 @@
 <?php
 
-
 namespace App\Form;
 
-
-use Symfony\Component\Form\AbstractType;
 use App\Entity\CustomersRating;
+use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\HiddenType;
-use Symfony\Component\Form\Extension\Core\Type\TextareaType;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
+use Symfony\Component\Form\Extension\Core\Type\TextareaType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
 class RateOfferType extends AbstractType
 {
-    public function buildForm(FormBuilderInterface $builder, array $options)
+    #[\Override]
+    public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
             ->add('rating', HiddenType::class, [
                 'label' => 'Your rating',
                 'label_attr' => [
-                    'class' => 'sr-only'
+                    'class' => 'sr-only',
                 ],
-                'data' => 3
+                'data' => 3,
             ])
             ->add('comment', TextareaType::class, [
                 'required' => false,
                 'label' => 'Comment',
                 'label_attr' => [
-                    'class' => 'sr-only'
+                    'class' => 'sr-only',
                 ],
                 'attr' => [
                     'class' => 'form-control',
                     'placeholder' => 'Comment',
-                    'rows' => 6
+                    'rows' => 6,
                 ],
             ])
             ->add('submit', SubmitType::class, [
-                'label' => 'Rate offer'
-            ]);
+                'label' => 'Rate offer',
+            ])
+        ;
     }
-    public function configureOptions(OptionsResolver $resolver)
+
+    #[\Override]
+    public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver->setDefaults([
-            'data_class' => CustomersRating::class
+            'data_class' => CustomersRating::class,
         ]);
     }
 }
