@@ -15,7 +15,8 @@ use Symfony\Component\OptionsResolver\OptionsResolver;
 
 class RegistrationType extends AbstractType
 {
-    public function buildForm(FormBuilderInterface $builder, array $options)
+    #[\Override]
+    public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
             ->add('firstName', TextType::class, [
@@ -23,29 +24,29 @@ class RegistrationType extends AbstractType
                 'label' => 'First Name',
                 'attr' => [
                     'class' => 'form-control',
-                    'autofocus' => true
-                ]
+                    'autofocus' => true,
+                ],
             ])
             ->add('lastName', TextType::class, [
                 'required' => true,
                 'label' => 'Last Name',
                 'attr' => [
-                    'class' => 'form-control'
-                ]
+                    'class' => 'form-control',
+                ],
             ])
             ->add('email', EmailType::class, [
                 'required' => true,
                 'label' => 'Email',
                 'attr' => [
-                    'class' => 'form-control'
-                ]
+                    'class' => 'form-control',
+                ],
             ])
             ->add('password', RepeatedType::class, [
                 'type' => PasswordType::class,
                 'invalid_message' => 'Passwords do not match.',
                 'options' => ['attr' => ['class' => 'form-control']],
                 'required' => true,
-                'first_options'  => ['label' => 'Password'],
+                'first_options' => ['label' => 'Password'],
                 'second_options' => ['label' => 'Repeat Password'],
             ])
             ->add('subscribeNewsletter', CheckboxType::class, [
@@ -53,21 +54,20 @@ class RegistrationType extends AbstractType
                 'required' => false,
                 'label' => 'Sign up for Newsletter',
                 'attr' => [
-                    'checked' => 1
-                ]
+                    'checked' => 1,
+                ],
             ])
             ->add('submit', SubmitType::class, [
                 'label' => 'Register',
                 'attr' => [
-                    'class' => 'btn btn-register'
-                ]
+                    'class' => 'btn btn-register',
+                ],
             ])
-
-
         ;
     }
 
-    public function configureOptions(OptionsResolver $resolver)
+    #[\Override]
+    public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver->setDefaults([
             'data_class' => User::class,
