@@ -1,81 +1,69 @@
 <?php
 
-namespace App\Security;
+declare(strict_types=1);
 
+namespace App\Security;
 
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
-use Symfony\Component\Security\Core\User\UserInterface;
-use Symfony\Component\Security\Core\Security;
-use Symfony\Component\Security\Core\User\UserProviderInterface;
-use Symfony\Component\Security\Guard\Authenticator\AbstractFormLoginAuthenticator;
-use App\Repository\UserRepository;
-use Symfony\Component\Routing\RouterInterface;
-use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
+use Symfony\Component\Security\Http\Authenticator\AbstractLoginFormAuthenticator;
+use Symfony\Component\Security\Http\Authenticator\Passport\Badge\CsrfTokenBadge;
+use Symfony\Component\Security\Http\Authenticator\Passport\Badge\RememberMeBadge;
+use Symfony\Component\Security\Http\Authenticator\Passport\Badge\UserBadge;
+use Symfony\Component\Security\Http\Authenticator\Passport\Credentials\PasswordCredentials;
+use Symfony\Component\Security\Http\Authenticator\Passport\Passport;
+use Symfony\Component\Security\Http\SecurityRequestAttributes;
 use Symfony\Component\Security\Http\Util\TargetPathTrait;
 
-class LoginFormAuthenticator extends AbstractFormLoginAuthenticator
+class LoginFormAuthenticator extends AbstractLoginFormAuthenticator
 {
     use TargetPathTrait;
-    private $userRepository;
-    private $router;
-    private $passwordEncoder;
 
-    public function __construct(UserRepository $userRepository, RouterInterface $router, UserPasswordEncoderInterface $passwordEncoder)
-    {
-        $this->userRepository = $userRepository;
-        $this->router = $router;
-        $this->passwordEncoder = $passwordEncoder;
+    public function __construct(
+        private readonly UrlGeneratorInterface $urlGenerator,
+    ) {
     }
 
-    public function supports(Request $request)
+    #[\Override]
+    public function authenticate(Request $request): Passport
     {
-        return $request->attributes->get('_route') === 'app_login'&& $request->isMethod('POST');
-    }
+        $email = $request->request->get('email', '');
+        \assert(\is_string($email));
 
-    public function getCredentials(Request $request)
-    {
-        //this is called immediately when supports returns true
-        //dump($request->request->all());die;
-        $postData = $request->request->get('login');
-        $email = $postData['email'];
-        $pass = $postData['password'];
-        $credentials = [
-            'email' => $email,
-            'password' => $pass
-        ];
+        $password = $request->request->get('password', '');
+        \assert(\is_string($password));
 
-        $request->getSession()->set(
-            Security::LAST_USERNAME,
-            $credentials['email']
+        $csrfToken = $request->request->get('_csrf_token');
+        \assert(\is_string($csrfToken));
+
+        $request->getSession()->set(SecurityRequestAttributes::LAST_USERNAME, $email);
+
+        return new Passport(
+            new UserBadge($email),
+            new PasswordCredentials($password),
+            [
+                new CsrfTokenBadge('authenticate', $csrfToken),
+                new RememberMeBadge(),
+            ]
         );
-        return $credentials;
     }
 
-    public function getUser($credentials, UserProviderInterface $userProvider)
+    #[\Override]
+    public function onAuthenticationSuccess(Request $request, TokenInterface $token, string $firewallName): Response
     {
-        return $this->userRepository->findOneBy(['email' => $credentials['email']]);
-    }
-
-    public function checkCredentials($credentials, UserInterface $user)
-    {
-        return $this->passwordEncoder->isPasswordValid($user, $credentials['password']);
-    }
-
-
-    public function onAuthenticationSuccess(Request $request, TokenInterface $token, $providerKey)
-    {
-        if($targetPath = $this->getTargetPath($request->getSession(), $providerKey))
-        {
+        if (($targetPath = $this->getTargetPath($request->getSession(), $firewallName)) !== null && ($targetPath = $this->getTargetPath($request->getSession(), $firewallName)) !== '' && ($targetPath = $this->getTargetPath($request->getSession(), $firewallName)) !== '0') {
             return new RedirectResponse($targetPath);
         }
 
-        return new RedirectResponse($this->router->generate('home'));
+        return new RedirectResponse($this->urlGenerator->generate('home'));
     }
 
-    protected function getLoginUrl()
+    #[\Override]
+    protected function getLoginUrl(Request $request): string
     {
-        return $this->router->generate('app_login');
+        return $this->urlGenerator->generate('app_login');
     }
 }
